chore: display basic information on endpoint view
